+++ src/app/locales/en/components/-browser-columns-name/filesystem.js
@@ -16,11 +16,10 @@ export default {
     modification: 'Last time the item\'s content was modified. For files, indicates the time of the last write or truncate operation on the file. For directories, updated when a child item is added, renamed or deleted.',
     atime: 'Last time the item was accessed.',
     ctime: 'Last time the item\'s metadata was changed, e.g. permissions. It\' s also always updated when the modification time changes.',
-    posixPermissions: 'File permission mask for UNIX or Linux systems for specifying access rights to files or directories.<br>Permissions for group refer to all space members.<br>Permissions for other are considered when a public share is accessed (as an anonymous guest).',
+    posixPermissions: 'Permissions regulating data access rights. It can be expressed either by the standard POSIX permissions system or by the Access Control List (ACL). If there is an ACL specified, the POSIX permissions are ignored.<br>Click on the cell to see more details.',
   },
   subname: {
     modification: 'content',
     ctime: 'metadata',
-    posixPermissions: 'POSIX'
   },
 };